Remove un-needed clones in tracking undo stack
With this change, we no longer need to clone the Editor's LineBuffer
on edits that don't cause a new insertion into the undo stack.

Previously, the EditStack always needed to have a copy of the current state.
With some tweaks to the API now the Editor has the sole copy of the
current state, so it can freely overwrite the LineBuffer when no new
undo point is needed.